feat: V2.6.0 架构重构与契约层建设
- 版本号统一更新为 2.6.0
- 新增 shared/ 契约层模块，实现前后端数据契约统一
- 数据飞轮插件系统完整实现（P4-4C 集成）
- 测试覆盖大幅增强，新增 300+ 单元测试
- 论文实验脚本与结果数据完整提交
- 修复多处 API 契约一致性问题
- 优化配置管理与环境变量处理

+++ python/app/ai/lnn/models/torch_cfc_model.py
@@ -66,7 +66,7 @@ class CFCLayer(nn.Module):
             nn.Linear(hidden_size, hidden_size),
         )
 
-        self.dt = nn.Parameter(torch.tensor(dt), requires_grad=False)
+        self.dt = nn.Parameter(torch.tensor(dt, dtype=torch.float32), requires_grad=False)
 
         self._init_weights()
